requirements.txt
anthropic
fastapi
uvicorn
python-dotenv
httpx
chromadb
slack-sdk
apscheduler
psycopg2-binary
sqlalchemy
